Clean-ups and improvements to TranslationDownloadActivity.

+++ src/net/zionsoft/obadiah/Utils.java
@@ -6,6 +6,8 @@ public class Utils
 {
     static public boolean removeDirectory(File directory)
     {
+        if (directory == null)
+            return true;
         File[] files = directory.listFiles();
         for (File file : files)
             file.delete();
